--- www/.gitignore
@@ -1,3 +1,0 @@
-/build
-roc_repl_wasm.js
-roc_repl_wasm_bg.wasm
