+++ www/content/design_goals.md
