+++ www/content/fast.md
