+++ www/content/functional.md
