+++ www/content/repl/index.md
@@ -1,0 +1,15 @@
+
+# The rockin’ Roc REPL
+
+<div id="repl-container" role="presentation">
+    <div id="repl" role="presentation">
+        <code class="history">
+          <div id="repl-intro-text">Enter an expression to evaluate, or a definition (like <span class="color-blue">x = 1</span>) to use later.</div>
+          <div id="history-text" aria-live="polite"></div>
+        </code>
+        <div id="repl-prompt" role="presentation">»</div>
+        <textarea aria-label="Input Roc code here, then press Enter to submit it to the REPL" rows="5" id="source-input" placeholder="Enter some Roc code here."></textarea>
+    </div>
+</div>
+<script type="module" src="/site.js"></script>
+</section>
