--- www/generate_tutorial/.gitignore
@@ -1,2 +1,0 @@
-tutorial
-src/output/tutorial.html
