--- www/public/homepage.css
@@ -1,15 +1,0 @@
-html {
-    font-family: sans-serif;
-    line-height: 145%;
-}
-
-li {
-    margin-bottom: 0.5rem;
-}
-
-@media only screen and (min-device-width: 900px) {
-    body {
-        width: 900px;
-        margin: 24px auto;
-    }
-}
