--- www/public/repl/index.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-
-<head>
-  <meta charset="utf-8">
-
-  <title>REPL</title>
-  <link rel="stylesheet" href="/site.css" />
-  <link rel="stylesheet" href="/repl/repl.css" />
-  <link rel="icon" href="/favicon.svg">
-  <!-- Safari ignores rel="icon" and only respects rel="mask-icon". It will render the SVG with
-    fill="#000" unless this `color` attribute here is hardcoded (not a CSS `var()`) to override it.
-  -->
-  <link rel="mask-icon" href="/favicon.svg" color="#7d59dd">
-</head>
-
-<body>
-  <div id="top-bar">
-    <nav><a id="nav-home-link" href="/" title="The Roc Programming Language">roc</a>
-      <div id="top-bar-links"><a href="/tutorial">tutorial</a><a
-          href="https://github.com/roc-lang/roc/tree/main/getting_started">install</a><a href="/repl">repl</a><a
-          href="/builtins">docs</a></div>
-    </nav>
-  </div>
-  <main id="repl">
-    <h1>The rockin’ Roc REPL</h1>
-    <code class="history">
-      <div id="help-text"></div>
-      <div id="history-text"><div id="loading-message">Loading REPL WebAssembly module…please wait!</div></div>
-    </code>
-    <div id="repl-prompt" role="presentation">»</div>
-    <textarea rows="5" id="source-input" placeholder="You can enter Roc code here once the REPL loads!"
-        disabled></textarea>
-  </main>
-  <script type="module" src="/repl/repl.js"></script>
-</body>
-
-</html>
