--- www/public/repl/repl.css
@@ -1,173 +1,0 @@
-main {
-  display: flex;
-  flex-direction: column;
-}
-
-h1 {
-  font-family: "Lato";
-  font-size: 48px;
-  line-height: 48px;
-  padding: 48px 0;
-  margin: 0;
-  color: var(--header-link-color);
-}
-
-#source-input-wrapper {
-  position: relative;
-  width: 100%;
-  box-sizing: border-box;
-}
-
-#source-input-wrapper::before {
-  content: "» ";
-  position: absolute;
-  left: 15px;
-  top: 18px;
-  line-height: 16px;
-  height: 16px;
-  z-index: 2;
-  font-family: var(--font-mono);
-  color: var(--cyan);
-  /* Let clicks pass through to the textarea */
-  pointer-events: none;
-}
-
-#source-input {
-  width: 100%;
-  font-family: var(--font-mono);
-  color: var(--code-color);
-  background-color: var(--code-bg);
-  display: inline-block;
-  height: 76px;
-  padding: 16px;
-  padding-left: 36px;
-  border: 1px solid transparent;
-  margin: 0;
-  margin-bottom: 2em;
-  box-sizing: border-box;
-}
-
-#source-input:focus {
-  border: 1px solid var(--cyan);
-  box-sizing: border-box;
-  outline: none;
-}
-
-li {
-  margin: 8px;
-}
-
-.history {
-  padding: 1em;
-  padding-bottom: 0;
-}
-
-#help-text,
-#history-text {
-  white-space: pre-wrap;
-
-}
-
-#history-text {
-  margin-top: 16px;
-}
-
-#loading-message {
-  text-align: center;
-  /* approximately match height after loading and printing help message */
-  height: 96px;
-}
-
-.history-item {
-  margin-bottom: 24px;
-}
-
-.history-item .input {
-  margin: 0;
-  margin-bottom: 8px;
-}
-
-.history-item .output {
-  margin: 0;
-}
-
-.panic {
-  color: red;
-}
-
-.input-line-prefix {
-  color: var(--cyan);
-}
-
-.color-red {
-  color: red;
-}
-
-.color-green {
-  color: var(--green);
-}
-
-.color-yellow {
-  color: var(--orange);
-}
-
-.color-blue {
-  color: var(--cyan);
-}
-
-.color-magenta {
-  color: var(--magenta);
-}
-
-.color-cyan {
-  color: var(--cyan);
-}
-
-.color-white {
-  /* Really this isn't white so much as "default text color." For the repl, this should be black
-  in a light color scheme, and only white in dark mode. The name could be better! */
-  color: black;
-}
-
-@media (prefers-color-scheme: dark) {
-  .color-white {
-    color: white;
-  }
-}
-
-.bold {
-  font-weight: bold;
-}
-
-.underline {
-  text-decoration: underline;
-}
-
-
-/* Mobile-friendly screen width */
-@media only screen and (max-width: 767px) {
-  h1 {
-    font-size: 24px !important;
-    margin: 0;
-    padding: 16px 0;
-    text-align: center;
-  }
-
-  #repl {
-    margin: 0;
-    padding: 0;
-    min-height: calc(100vh - var(--top-bar-height));
-  }
-
-  code.history {
-    flex-grow: 1;
-  }
-
-  #source-input {
-    margin: 0
-  }
-
-  #loading-message {
-    margin: 0;
-  }
-}
